Add git-remote-gcrypt
Add `gcrypt::` in front of a remote to encrypt it.

Your default GPG key will be used, and that can be set in `.gnupg/gpg.conf`
using `default-key <email>`.

The `remote.<name>.gcrypt-participants` and `gcrypt.participants` git options
set the GnuPG key to use for a specific remote, or by default.